Rename helper to json_helper
